requirements.txt
# ═══════════════════════════════════════════════════════════════
# CONFLUENCE MCP SERVER - REQUIREMENTS
# ═══════════════════════════════════════════════════════════════

# Core MCP Framework
fastmcp>=0.3.0

# Vector Database
qdrant-client>=1.7.0

# Embeddings (GPU accelerated)
fastembed>=0.2.0
flashrank>=0.2.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# GPU Support (CUDA 12.4)
torch>=2.1.0
onnxruntime-gpu>=1.16.0

# HTTP Client
httpx>=0.25.0

# HTML Parsing
beautifulsoup4>=4.12.0
markdownify>=0.11.0

# Configuration & Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Image Processing (for diagram search)
Pillow>=10.0.0

# Text Processing (for extractive summaries)
nltk>=3.8.0
scikit-learn>=1.3.0

# Async Support
aiofiles>=23.0.0

# Caching
cachetools>=5.3.0

# Logging & Observability
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
